refactor(backend): extract ogie data→RawProductPage helpers to ogieRawProductPage
